// src/Types/user.ts
export interface User {
    prefix_name: string;
    u_id: number;
    prefix_id: number;
    u_fname: string;
    u_lname: string;
    age: string;
    gender: string;
    level_id: number;
    u_id_card: string;
    position_id: number;
    ex_position_id: number;
    type_p_id: number;
    course_id: number;
    branch_id: number;
    salary: string;
    work_start: string;
    u_tel: string;
    u_img: string | File;
    u_email: string;
    u_pass: string;
    ex_position_name: string;
    level_name: string;
    position_name: string;
    type_p_name: string;
    course_name: string;
    branch_name: string;
}